add youtube channel url, update carousel images mock
