+++ src/utils/urls.ts
@@ -1,12 +1,11 @@
 export enum AppUrls {
-  Home = '/',
-  Media = '/cchv-tv',
-  News = '/noticias/',
-  NewsPost = '/noticias/',
-  Programs = '/programas',
-  About = '/quienes-somos',
-  AboutTeam = '/quienes-somos/equipo',
-  AboutTransparency = '/quienes-somos/transparencia'
+  Home = "/",
+  News = "/noticias/",
+  NewsPost = "/noticias/",
+  Programs = "/programas",
+  About = "/quienes-somos",
+  AboutTeam = "/quienes-somos/equipo",
+  AboutTransparency = "/quienes-somos/transparencia",
 }
 
-export default AppUrls 
+export default AppUrls
